enhanced moved into client; client renaming
