--- include/enhanced.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <WiFiClient.h>
-
-int pushEnhClient(WiFiClient* client, uint8_t c, uint8_t d, bool log);
-void handleEnhClient(WiFiClient* client);
